rename file
https://coveord.atlassian.net/browse/KIT-633
